+++ src/validation/document-definition-schema.js
@@ -54,7 +54,8 @@ module.exports = exports = joi.object().options({ convert: false }).keys({
               otherwise: integerSchema.min(1)
             })),
         supportedExtensions: dynamicConstraintSchema(joi.array().min(1).items(joi.string())),
-        supportedContentTypes: dynamicConstraintSchema(joi.array().min(1).items(nonEmptyStringSchema))
+        supportedContentTypes: dynamicConstraintSchema(joi.array().min(1).items(nonEmptyStringSchema)),
+        filenameRegexPattern: dynamicConstraintSchema(joi.object().type(RegExp))
       })),
 
   channels: dynamicConstraintSchema(