+++ test/resources/attachment-constraints-doc-definitions.js
@@ -32,6 +32,15 @@
       }
     }
   },
+  staticAttachmentFilenameRegexPatternDoc: {
+    typeFilter: simpleTypeFilter,
+    channels: { write: 'write' },
+    allowAttachments: true,
+    attachmentConstraints: {
+      filenameRegexPattern: /^(foo|bar)\.(xls|xlsx)$/
+    },
+    propertyValidators: { }
+  },
   dynamicAttachmentsDoc: {
     typeFilter: simpleTypeFilter,
     channels: { write: 'write' },
@@ -57,6 +66,9 @@
         },
         requireAttachmentReferences: function(doc, oldDoc) {
           return doc.requireAttachmentReferences;
+        },
+        filenameRegexPattern: function(doc, oldDoc) {
+          return doc.filenameRegexPattern ? new RegExp(doc.filenameRegexPattern) : null;
         }
       };
     },
@@ -87,6 +99,9 @@
         arrayElementsValidator: {
           type: 'attachmentReference'
         }
+      },
+      filenameRegexPattern: {
+        type: 'string'
       }
     }
   }