+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/). All notable c
 ## [Unreleased]
 ### Added
 - [#311](https://github.com/Kashoo/synctos/issues/311): Case insensitive equality constraint for strings
+- [#313](https://github.com/Kashoo/synctos/issues/313): Attachment filename regular expression constraint
 
 ## [2.4.0] - 2018-04-24
 ### Added